+++ plugins/org.eclipse.osee.framework.ui.branch.graph/src/org/eclipse/osee/framework/ui/branch/graph/model/GraphLoader.java
@@ -151,7 +151,7 @@ public class GraphLoader {
             BranchId branch = TokenFactory.createBranch(chStmt.getLong("branch_id"));
             TxData txData = new TxData(branch, chStmt.getInt("author"), chStmt.getTimestamp("time"),
                chStmt.getString("osee_comment"), chStmt.getInt("tx_type"), chStmt.getInt("commit_art_id"),
-               chStmt.getInt("transaction_id"));
+               chStmt.getLong("transaction_id"));
             txDatas.add(txData);
          }
       } finally {
