+++ plugins/org.eclipse.osee.orcs/src/org/eclipse/osee/orcs/search/TxQueryBuilder.java
@@ -26,9 +26,9 @@ public interface TxQueryBuilder<T> {
 
    T andTxId(TransactionId id);
 
-   T andTxId(int... id) throws OseeCoreException;
+   T andTxId(long... id) throws OseeCoreException;
 
-   T andTxIds(Collection<Integer> ids) throws OseeCoreException;
+   T andTxIds(Collection<Long> ids) throws OseeCoreException;
 
    T andTxId(Operator op, int id) throws OseeCoreException;
